+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * [x] [GameAPI](src/infrastructure/game_api/api.py)
 * [x] LanguageTool и словарь исключений
 * [x] Настроить Docker-конфигурацию
-* [ ] [LogFile](src/infrastructure/filesystem.py) с ротацией
+* [x] [LogFile](src/infrastructure/filesystem/log_reader/log_file.py) с ротацией
 * [ ] Диспетчер логов
 * [ ] Маршруты для вызова use cases
 * [ ] PostgreSQL